♻ refactor : [공통으로 사용하는 컴포넌트 파일 생성] | [김진영] | [ColorBackgroundWithText.kt] - 기존 에그머니나 페이지 수정
