+++ android/app/src/main/java/com/shbhack/eggmoneyna/ui/common/component/ColorBackgroundWithText.kt
@@ -1,4 +1,4 @@
-package com.shbhack.eggmoneyna.ui.eggmoneyna
+package com.shbhack.eggmoneyna.ui.common.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -18,21 +19,22 @@ import com.shbhack.eggmoneyna.R
 import com.shbhack.eggmoneyna.ui.theme.EggmoneynaPurple
 
 @Composable
-fun ColorBackgroundWithText() {
+fun ColorBackgroundWithText(color: Color, title: String, content: String) {
     Column(
         modifier = Modifier
             .fillMaxWidth()
-            .background(color = EggmoneynaPurple)
+            .wrapContentHeight()
+            .background(color = color)
             .padding(32.dp)
     ) {
         Text(
-            text = stringResource(id = R.string.eggmoneyna_rest_money_title), fontWeight = FontWeight.Medium, fontSize = 18.sp, color = Color.White
+            text = title, fontWeight = FontWeight.Medium, fontSize = 18.sp, color = Color.White
         )
         Spacer(
             modifier = Modifier.size(4.dp)
         )
         Text(
-            text = "500,000원", fontWeight = FontWeight.SemiBold,
+            text = content, fontWeight = FontWeight.SemiBold,
             fontSize = 28.sp, color = Color.White
         )
     }
